+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to
 - 🐛(backend) manage ole2 compound document format
 - ♻️(backend) increase user short_name field length
 - 🐛(docker) avoid downloading mime.types at build time (CI stability)
+- 🐛(front) fix explorer tree create folder action labels
 
 ### Security